backend/app/requirements.txt
# FastAPI + Server
fastapi==0.110.0
uvicorn[standard]==0.29.0

# Database
sqlalchemy[asyncio]==2.0.29
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Pydantic for config/schema validation
pydantic==2.7.1

# AI / ML Utils
scikit-learn==1.5.0
prophet==1.1.5
statsmodels==0.14.2
numpy==1.26.4
pandas==2.2.2

# Cloud SDK
boto3==1.34.84

# Monitoring
prometheus-fastapi-instrumentator==7.0.0

# Utils
requests==2.32.3
python-dotenv==1.0.1
pyyaml==6.0.2
